Move CursorOptions up into it's own file
