+++ src/renderer/dx/DxRenderer.cpp
@@ -1298,6 +1298,7 @@ try
                            _dwriteFactory.Get(),
                            spacing,
                            _glyphCell,
+                           {},
                            D2D1_DRAW_TEXT_OPTIONS_ENABLE_COLOR_FONT);
 
     // Layout then render the text
@@ -1427,7 +1428,7 @@ enum class CursorPaintType
 // - options - Packed options relevant to how to draw the cursor
 // Return Value:
 // - S_OK or relevant DirectX error.
-[[nodiscard]] HRESULT DxEngine::PaintCursor(const IRenderEngine::CursorOptions& options) noexcept
+[[nodiscard]] HRESULT DxEngine::PaintCursor(const CursorOptions& options) noexcept
 try
 {
     // if the cursor is off, do nothing - it should not be visible.
